refactor: extract optimistic event handling to separate functions
